pyproject.toml
[project]
name = "ai-qa-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = ">=3.12,<4.0"          # add the <4.0 upper-bound
openai = "*"
langchain = "*"
chromadb = "*"
fastapi = "0.115.9"
uvicorn = "*"
python-dotenv = "*"
pytest = "*"
